fix: update to latest sdk and fix issues with large discount
