+++ ava.config.cjs
@@ -1,0 +1,1 @@
+module.exports = require("near-runner-ava/ava.config.cjs")
