+++ contracts/contract-utils/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "contract-utils"
+version = "0.1.0"
+authors = ["Willem Wyndham <willem@cs.umd.edu>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+near-sdk = "4.0.0-pre.2"
