+++ contracts/sandbox-linkdrop/src/lib.rs
@@ -1,10 +1,7 @@
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
-use near_sdk::json_types::{U128};
-use near_sdk::{
-    env, ext_contract, near_bindgen, AccountId, 
-    Promise, PromiseResult, PublicKey, Gas
-};
-
+use near_sdk::json_types::U128;
+use near_sdk::{AccountId, Gas, Promise, PublicKey, env, ext_contract, near_bindgen};
+use contract_utils::is_promise_success;
 
 /// Gas attached to the callback from account creation.
 pub const ON_CREATE_ACCOUNT_CALLBACK_GAS: Gas = Gas(20_000_000_000_000);
@@ -24,58 +21,46 @@ pub trait ExtLinkDrop {
     /// Callback after creating account and claiming linkdrop.
     fn on_account_created_and_claimed(&mut self, amount: U128) -> bool;
 }
-fn is_promise_success() -> bool {
-  assert_eq!(
-      env::promise_results_count(),
-      1,
-      "Contract expected a result on the callback"
-  );
-  match env::promise_result(0) {
-      PromiseResult::Successful(_) => true,
-      _ => false,
-  }
-}
-
 
 #[near_bindgen]
 impl LinkDrop {
-   /// Create new account without linkdrop and deposit passed funds (used for creating sub accounts directly).
-   #[payable]
-   pub fn create_account(
-       &mut self,
-       new_account_id: AccountId,
-       new_public_key: PublicKey,
-   ) -> Promise {
-       assert!(
-           env::is_valid_account_id(new_account_id.as_bytes()),
-           "Invalid account id"
-       );
-       let amount = env::attached_deposit();
-       Promise::new(new_account_id)
-           .create_account()
-           .add_full_access_key(new_public_key.into())
-           .transfer(amount)
-           .then(ext_self::on_account_created(
-               env::predecessor_account_id(),
-               amount.into(),
-               &env::current_account_id(),
-               NO_DEPOSIT,
-               ON_CREATE_ACCOUNT_CALLBACK_GAS,
-           ))
-   }
+    /// Create new account without linkdrop and deposit passed funds (used for creating sub accounts directly).
+    #[payable]
+    pub fn create_account(
+        &mut self,
+        new_account_id: AccountId,
+        new_public_key: PublicKey,
+    ) -> Promise {
+        assert!(
+            env::is_valid_account_id(new_account_id.as_bytes()),
+            "Invalid account id"
+        );
+        let amount = env::attached_deposit();
+        Promise::new(new_account_id)
+            .create_account()
+            .add_full_access_key(new_public_key.into())
+            .transfer(amount)
+            .then(ext_self::on_account_created(
+                env::predecessor_account_id(),
+                amount.into(),
+                env::current_account_id(),
+                NO_DEPOSIT,
+                ON_CREATE_ACCOUNT_CALLBACK_GAS,
+            ))
+    }
 
-   /// Callback after executing `create_account`.
-   pub fn on_account_created(&mut self, predecessor_account_id: AccountId, amount: U128) -> bool {
-       assert_eq!(
-           env::predecessor_account_id(),
-           env::current_account_id(),
-           "Callback can only be called from the contract"
-       );
-       let creation_succeeded = is_promise_success();
-       if !creation_succeeded {
-           // In case of failure, send funds back.
-           Promise::new(predecessor_account_id).transfer(amount.into());
-       }
-       creation_succeeded
-   }
+    /// Callback after executing `create_account`.
+    pub fn on_account_created(&mut self, predecessor_account_id: AccountId, amount: U128) -> bool {
+        assert_eq!(
+            env::predecessor_account_id(),
+            env::current_account_id(),
+            "Callback can only be called from the contract"
+        );
+        let creation_succeeded = is_promise_success(None);
+        if !creation_succeeded {
+            // In case of failure, send funds back.
+            Promise::new(predecessor_account_id).transfer(amount.into());
+        }
+        creation_succeeded
+    }
 }
