+++ contracts/tenk/Cargo.toml
@@ -8,9 +8,10 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-near-sdk = "4.0.0-pre.1"
-near-contract-standards = "4.0.0-pre.1"
+near-sdk = "4.0.0-pre.2"
+near-contract-standards = "4.0.0-pre.2"
 serde = "1.0.0"
+contract-utils =  { path = "../contract-utils" }
 
 [dev-dependencies]
 rand = "0.7.2"
