+++ package.json
@@ -5,16 +5,16 @@
   "scripts": {
     "pretest": "yarn build",
     "build": "cargo build --target wasm32-unknown-unknown --release",
-    "test": "near-runner-jest",
+    "test": "near-runner-ava",
     "test:testnet": "NEAR_RUNNER_NETWORK=testnet yarn test",
     "test:unit": "cargo test",
-    "test:ci": "yarn test:unit && yarn test && yarn test:testnet"
+    "test:ci": "yarn test:unit && yarn test"
   },
   "devDependencies": {
     "acquit": "^1.2.1",
     "acquit-markdown": "^0.1.0",
     "eslint": "^7.29.0",
-    "near-runner-jest": "^0.1.8",
+    "near-runner-ava": "^0.1.0",
     "near-units": "^0.1.9"
   }
 }
